# Optimize mocking