--- a/rebar.config
+++ b/rebar.config
@@ -5,8 +5,13 @@
    {git, "https://github.com/uwiger/parse_trans.git", {ref, "64b9b77"}}},
   {eunit_addons, "1.0.1"}
  ]}.
+
+{edoc_opts,
+ [{preprocess, true}]}.
+
 {profiles,
  %% Add the edown dependency (for generating markdown docs) when
  %% running with the edown profile ==> make it an optional dependency.
  [{edown, [{deps, [edown]},
-           {edoc_opts, [{doclet, edown_doclet}]}]}]}.
+           {edoc_opts, [{doclet, edown_doclet},
+                        {preprocess, true}]}]}]}.