This is a set of commits that improves the performance when mocking by approximately 50%. I also have some questions towards the end of the text.

For a fairly large unit test module with 44 test cases with many mocked functions, the time to execute all the tests was reduced by 52% as indicated in the table below where upstream/master is the first line, and the HEAD is the last line. Lower time is better.

Execution time | Commit
-----|------------------------------------------------
67s  | Baseline (the b2797f7 _Bump version to 1.2.0_)
...  | Prepare the cache tab to hold more types of items
...  | Collect unload-info during mocking
60s  | Load lists of modules at a time
55s  | Mock modules in parallel
47s  | Reload modules instead of unloading them
32s  | Let renamed modules stay loaded between sessions

The unit tests mocked some 70 functions across 20 modules of varying sizes, ranging from 3.6kB to 6.2MB. This was the file size of the beam files, which were compiled with debug_info. Erlang 23. I have not measured with Erlang 24 or any other Erlang version.

On the smaller scale it seems the optimization gives roughly the same performance gain. Below are some results (before ⟶ after).

Execution time | Mocking stats
---------------|--------------------
158ms ⟶ 79ms | 2 functions in 2 modules mocked, beam file sizes: 44kB and 36kB
112ms ⟶ 66ms | 1 function mocked, beam file size was 8kB

In the original 67 second run of the big unit test of 44 test cases, the time was originally spent roughly as follows (only major time contributors are shown):

Time | Step
-----|--------------------
54s  | `mock_and_load_mods/1`
23s  | ├ `reload_mod_under_different_name/3`
14s  | ├ `get_exported_fas_and_object_code/1`
6s   | └ `mk_mocking_mod/3` (results were cached, so this took time only once)
8s   | `unload_mods/1` (in `end_session`)

----

The general approach of the commits is to minimize unnecessary loads and unloads of modules, since these can be costly in modern Erlang (since unloading implies garbage collection of all processes to copy data from the module's literal area.)

To minimize unnecessary loading, it now keeps track also of the original modules in the ets cache table. Previously only the mockings were cached.

Another change is that instead of iterating over modules, it passes a list of modules to `code` for loading, since this was somewhat faster at times (although when I measured, this was very jittery for reasons I didn't see). I also tried to exploit cpu cores where available, for some parts of the processing.

----

I have a question about desired Erlang version compatibility: The code now uses [`code:atomic_load/1`](https://erlang.org/doc/man/code.html#atomic_load-1) and [`code:ensure_modules_loaded/1`](https://erlang.org/doc/man/code.html#ensure_modules_loaded-1) which were introduced in Erlang 19. So it means the mockgyver now requires at least Erlang 19.

Do you want to support Erlang versions
* only ≥ 21? If so, I can drop the [`?with_stacktrace`](https://github.com/tomas-abrahamsson/mockgyver/blob/optimize-mocking/src/mockgyver.erl#L397-L406) compatibility macro. **Edit2:** The [addition of `{preprocess,true}` to rebar.config](https://github.com/tomas-abrahamsson/mockgyver/blob/ba350c7/rebar.config) could then be dropped (2 places).
* ≥ 19? If so one could drop an Erlang 17-ism in [`get_module_checksum/1`](https://github.com/klajo/mockgyver/blob/b2797f7/src/mockgyver.erl#L1304-L1309) although of course this would not impact performance in any way.
* also pre 19? If so, I could add some kind of fallback to process modules individually instead of calling `code:ensure_modules_loaded/1` and `code:atomic_load/1`.

Another question: Mockgyver now uses `file:read_file/1` and [`beam_lib:md5(Filename)`](https://erlang.org/doc/man/beam_lib.html#md5-1) to reduce code loading. I think this means that mockgyver might perhaps not work well with embedded code loading. I don't know if it did earlier, or if it is important to make sure it does support this?

----

**Edit:** Forgot to say that I have tested the set of commits on a large project as well.

**Edit2:** Also forgot to say that review of the code is welcome, of course.