@tomas-abrahamsson 2021-07-24
I have an idea for improving performance even further. No code yet, only an idea so far. But I think it will require some changes to the state machine logic. So once I have some code, I think it will depend on #19 (or something similar).