@tomas-abrahamsson 2021-07-24
I added two commits early in the chain, because I think they made more sense there:
* Use the checksum of the original module rather than the renamed as cache for the mocking module.
* Introduce the term 'renamed' throughout (and corrected a few 'mocked' → 'mocking')

I also took the liberty to rebase on top of the updated master, assuming you want a linear history rather than a merge. (easy to undo if this was a premature guess from my side, just let me know)

Returning to your earlier topics:

> * Maybe we can have a chat to reason a bit about this...

Yes, absolutely!

> * embedded code loading:  ... What is it that makes it malfunction with embedded code loading? Won't it respect code paths and loading in the same way as dynamic code loading, only that it's done at the start, triggered by the boot script?

(guess this is a minor thing) Yes, maybe there is nothing that makes it malfunction. I was tracing the calls through `code` and the `code_server` into `erl_prim_file` and didn't manage to muster all the troups to compare this call chain to what the `file` and the `file_server` does for embedded mode and how `prim_file` and `erl_prim_file` relate to each other.